// src/lib/api/generated/main/models/aPIResponseLoginResponse.ts
/**
 * Generated by orval v7.5.0 🍺
 * Do not edit manually.
 * Devlens Admin API 문서
 * Devlens Admin API 문서
 * OpenAPI spec version: 1.0
 */
import type { LoginResponse } from './loginResponse';

export interface APIResponseLoginResponse {
  code?: number;
  message?: string;
  data?: LoginResponse;
}